requirements.txt
flask
flask-sqlalchemy
twilio
psycopg2-binary
python-dotenv
firebase-admin
requests
